[#139414339] Finish GCP integration test suite

+++ integration_tests/gcp/gcp_client_api_test.go
@@ -7,6 +7,7 @@ import (
 	"io"
 	"io/ioutil"
 	"os"
+	"time"
 
 	"golang.org/x/oauth2/google"
 	compute "google.golang.org/api/compute/v1"
@@ -23,6 +24,7 @@ var _ = Describe("GCPCLientAPI", func() {
 		var zone = os.Getenv("GCP_ZONE")
 		var credFile *os.File
 		var gcpClientAPI *GCPClientAPI
+		var instanceNameGUID, err = newUUID()
 		BeforeEach(func() {
 			var err error
 			credFile, err = ioutil.TempFile("/tmp", "gcpCred")
@@ -37,53 +39,100 @@ var _ = Describe("GCPCLientAPI", func() {
 				ConfigProjectName(project),
 			)
 			Expect(err).ShouldNot(HaveOccurred())
-			//createVM(instanceNameGUID)
+			createVM(instanceNameGUID, project, zone)
 		})
 		AfterEach(func() {
+			if instanceExists(instanceNameGUID, project, zone) {
+				deleteVM(instanceNameGUID, project, zone)
+			}
+
+			if diskExists(instanceNameGUID, project, zone) {
+				deleteDisk(instanceNameGUID, project, zone)
+			}
+
 			os.Remove(credFile.Name())
-			//deleteVM(instanceNameGUID)
 		})
 		Context("when calling CreateVM with valid arguments", func() {
-			var instanceNameGUID string
+			var instanceNameGUIDCreate string
 			var computeInstance compute.Instance
 			BeforeEach(func() {
 				var err error
-				instanceNameGUID, err = newUUID()
+				instanceNameGUIDCreate, err = newUUID()
 				Expect(err).ShouldNot(HaveOccurred())
-				Expect(instanceExists(instanceNameGUID, project, zone)).Should(BeFalse())
-				computeInstance = newComputeInstance(instanceNameGUID, zone, project)
+				Expect(instanceExists(instanceNameGUIDCreate, project, zone)).Should(BeFalse())
+				computeInstance = newComputeInstance(instanceNameGUIDCreate, project, zone)
 				err = gcpClientAPI.CreateVM(computeInstance)
 				Expect(err).ShouldNot(HaveOccurred())
 			})
 			AfterEach(func() {
-				deleteVM(instanceNameGUID, project, zone)
+				deleteVM(instanceNameGUIDCreate, project, zone)
 			})
 			It("then a new instance should have been created in GCP", func() {
-				Expect(instanceExists(instanceNameGUID, project, zone)).Should(BeTrue())
+				Expect(instanceExists(instanceNameGUIDCreate, project, zone)).Should(BeTrue())
 			})
 		})
 
-		XContext("when calling DeleteVM with valid arguments for a running instance", func() {
+		Context("when calling DeleteVM with valid arguments for a running instance", func() {
 			It("then the specified instance should not longer exist in GCP", func() {
-				Expect(true).Should(BeFalse())
+				Expect(instanceExists(instanceNameGUID, project, zone)).Should(BeTrue())
+				err = gcpClientAPI.DeleteVM(instanceNameGUID)
+				Expect(err).ShouldNot(HaveOccurred())
+				waitForDelete(instanceNameGUID, project, zone, instanceExists)
+				Expect(instanceExists(instanceNameGUID, project, zone)).Should(BeFalse())
 			})
 		})
 
-		XContext("when calling GetVMInfo with valid arguments for a running instance", func() {
-			It("then we should recieve all info about the matching instance in GCP", func() {
-				Expect(true).Should(BeFalse())
+		Context("when calling GetVMInfo with valid arguments for a running instance", func() {
+			var controlTag = "cliaas"
+			It("then we should recieve all info about the matching instance (by name) in GCP", func() {
+				Expect(instanceExists(instanceNameGUID, project, zone)).Should(BeTrue())
+				Expect(instanceStopped(instanceNameGUID, project, zone)).Should(BeFalse())
+				instance, err := gcpClientAPI.GetVMInfo(Filter{
+					NameRegexString: instanceNameGUID,
+					TagRegexString:  "",
+				})
+				Expect(err).ShouldNot(HaveOccurred())
+				Expect(instanceExists(instanceNameGUID, project, zone)).Should(BeTrue())
+				Expect(instanceStopped(instanceNameGUID, project, zone)).Should(BeFalse())
+				Expect(instance.Name).Should(Equal(instanceNameGUID))
+				Expect(instance.Tags).ShouldNot(BeNil())
+				Expect(instance.Tags.Items).Should(HaveLen(1))
+				Expect(instance.Tags.Items[0]).Should(Equal(controlTag))
+			})
+
+			It("then we should receive all info about the matching instance (by tag) in GCP", func() {
+				Expect(instanceExists(instanceNameGUID, project, zone)).Should(BeTrue())
+				Expect(instanceStopped(instanceNameGUID, project, zone)).Should(BeFalse())
+				instance, err := gcpClientAPI.GetVMInfo(Filter{
+					NameRegexString: "",
+					TagRegexString:  controlTag,
+				})
+				Expect(err).ShouldNot(HaveOccurred())
+				Expect(instanceExists(instanceNameGUID, project, zone)).Should(BeTrue())
+				Expect(instanceStopped(instanceNameGUID, project, zone)).Should(BeFalse())
+				Expect(instance.Name).Should(Equal(instanceNameGUID))
+				Expect(instance.Tags).ShouldNot(BeNil())
+				Expect(instance.Tags.Items).Should(HaveLen(1))
+				Expect(instance.Tags.Items[0]).Should(Equal(controlTag))
 			})
 		})
 
-		XContext("when calling StopVM with valid arguments for a running instance", func() {
+		Context("when calling StopVM with valid arguments for a running instance", func() {
 			It("then the specified instance should have been stopped in GCP", func() {
-				Expect(true).Should(BeFalse())
+				Expect(instanceExists(instanceNameGUID, project, zone)).Should(BeTrue())
+				Expect(instanceStopped(instanceNameGUID, project, zone)).Should(BeFalse())
+				err = gcpClientAPI.StopVM(instanceNameGUID)
+				Expect(err).ShouldNot(HaveOccurred())
+				Expect(instanceExists(instanceNameGUID, project, zone)).Should(BeTrue(), "does the instance exist?")
+				Eventually(func() bool {
+					return instanceStopped(instanceNameGUID, project, zone)
+				}, 30, 5).Should(BeTrue(), "is the instance stopped?")
 			})
 		})
 	})
 })
 
-func instanceExists(instanceNameGUID string, project string, zone string) bool {
+func getCompute() *compute.Service {
 	ctx := context.Background()
 	hc, err := google.DefaultClient(ctx, compute.CloudPlatformScope)
 	if err != nil {
@@ -94,8 +143,36 @@ func instanceExists(instanceNameGUID string, project string, zone string) bool {
 		Expect(err).ShouldNot(HaveOccurred())
 	}
 
-	call := c.Instances.List(project, zone)
+	return c
+}
+
+func deleteDisk(instanceNameGUID string, project string, zone string) {
+	c := getCompute()
+
+	call := c.Disks.Delete(project, zone, instanceNameGUID)
+	_, err := call.Do()
+	Expect(err).ShouldNot(HaveOccurred())
+
+	waitForDelete(instanceNameGUID, project, zone, diskExists)
+}
+
+func waitForDelete(instanceNameGUID string, project string, zone string, exists func(string, string, string) bool) {
+	var waitSeconds time.Duration = 10 * time.Second
+
+	for {
+		if !exists(instanceNameGUID, project, zone) {
+			break
+		}
+
+		time.Sleep(waitSeconds)
+	}
+}
+
+func diskExists(instanceNameGUID string, project string, zone string) bool {
+	c := getCompute()
+	call := c.Disks.List(project, zone)
 	list, err := call.Do()
+	Expect(err).ShouldNot(HaveOccurred())
 
 	for _, item := range list.Items {
 		if item.Name == instanceNameGUID {
@@ -105,8 +182,38 @@ func instanceExists(instanceNameGUID string, project string, zone string) bool {
 	return false
 }
 
-func newComputeInstance(instanceNameGUID string, zone string, project string) compute.Instance {
-	machineType := "zones/" + zone + "/machineTypes/n1-standard-1"
+func instanceExists(instanceNameGUID string, project string, zone string) bool {
+	c := getCompute()
+	call := c.Instances.List(project, zone)
+	list, err := call.Do()
+	Expect(err).ShouldNot(HaveOccurred())
+
+	for _, item := range list.Items {
+		if item.Name == instanceNameGUID {
+			return true
+		}
+	}
+
+	return false
+}
+
+func instanceStopped(instanceNameGUID string, project string, zone string) bool {
+	c := getCompute()
+	call := c.Instances.List(project, zone)
+	list, err := call.Do()
+	Expect(err).ShouldNot(HaveOccurred())
+
+	for _, item := range list.Items {
+		if item.Name == instanceNameGUID && (item.Status == "STOPPED" || item.Status == "STOPPING") {
+			return true
+		}
+	}
+
+	return false
+}
+
+func newComputeInstance(instanceNameGUID string, project string, zone string) compute.Instance {
+	machineType := "zones/" + zone + "/machineTypes/f1-micro"
 	nic := []*compute.NetworkInterface{
 		&compute.NetworkInterface{
 			Kind:    "compute#networkInterface",
@@ -114,7 +221,6 @@ func newComputeInstance(instanceNameGUID string, zone string, project string) co
 			Network: "global/networks/default",
 		},
 	}
-
 	sourceImage := "projects/debian-cloud/global/images/family/debian-8"
 	disks := []*compute.AttachedDisk{
 		&compute.AttachedDisk{
@@ -124,34 +230,41 @@ func newComputeInstance(instanceNameGUID string, zone string, project string) co
 			},
 		},
 	}
+	tags := &compute.Tags{
+		Items: []string{
+			"cliaas",
+		},
+	}
 
 	return compute.Instance{
 		Name:              instanceNameGUID,
 		MachineType:       machineType,
 		NetworkInterfaces: nic,
 		Disks:             disks,
+		Tags:              tags,
 	}
 
 }
 
-func createVM(instanceNameGUID string) {
+func createVM(instanceNameGUID string, project string, zone string) {
+	instance := newComputeInstance(instanceNameGUID, project, zone)
+
+	c := getCompute()
+
+	call := c.Instances.Insert(project, zone, &instance)
+	_, err := call.Do()
+	Expect(err).ShouldNot(HaveOccurred())
 
 }
 
 func deleteVM(instanceNameGUID string, project string, zone string) {
-	ctx := context.Background()
-	hc, err := google.DefaultClient(ctx, compute.CloudPlatformScope)
-	if err != nil {
-		Expect(err).ShouldNot(HaveOccurred())
-	}
-	c, err := compute.New(hc)
-	if err != nil {
-		Expect(err).ShouldNot(HaveOccurred())
-	}
+	c := getCompute()
 
 	call := c.Instances.Delete(project, zone, instanceNameGUID)
-	_, err = call.Do()
+	_, err := call.Do()
 	Expect(err).ShouldNot(HaveOccurred())
+
+	waitForDelete(instanceNameGUID, project, zone, instanceExists)
 }
 
 // newUUID generates a random UUID according to RFC 4122
